LSPS-9109 PDF Viewer - total pages invisible in IE 11
git js: 054dff665befde4a5afb0f955b6689d8473ca846

Number of total pages was invisible in IE11. The direct cause was that _wasLocalized property of the toolbar was never set to true. This was caused but exception in file validation - file is null in our case and IE11 throws an exception on new URL(null, ...).